Refactor scheduling and settings pages; remove unused components and update imports for consistency
