+++ royal-care-frontend/src/pages/SchedulingPage/SchedulingPage.module.css
@@ -1,0 +1,1 @@
+/* Place your SchedulingPage specific styles here */
