+++ royal-care-frontend/src/pages/SettingsDataPage/SettingsDataPage.module.css
@@ -1,0 +1,1 @@
+/* Place your SettingsDataPage specific styles here */
