+++ royal-care-frontend/src/pages/SettingsPage/SettingsPage.module.css
@@ -1,0 +1,1 @@
+/* Place your SettingsPage specific styles here */
